Implemented first minimal version of DSL parser and Simulator logic with general archtitecture to extend with new ops per device backend.
This was originally a C++ port, so i am getting adjusted to Rust its proc macros and syntax a bit. I will have to say that its "metaprogramming" equivalent is 10x better than in C++, it much logical, like parsing an AST tree.

My apologies if the rust code is not the prettiest, i am not super comfortable with Rust yet, using C/C++ and ASM mostly.

Also made some edits to the DSL interop API for fetching variables, as the Graph is created on compile time, but the nodes are run time objects, so you dont know the type of the variable.

I am really liking Rust so far, it codes so much faster and has is easier to get something with multiple packages going.